backend/requirements.txt
gradio==3.50.2
flask==2.3.3
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
tensorflow==2.13.0
joblib==1.3.2
flask-cors==4.0.0
gunicorn==21.2.0
python-dotenv==1.0.0
flask-limiter==3.5.0
psutil==7.0.0
requests==2.31.0
pytest==7.4.3
pytest-cov==4.1.0
pytest-html==4.1.1
pytest-mock==3.12.0
pytest-xdist==3.3.1
psycopg2-binary==2.9.7
sqlalchemy==2.0.21
flask-sqlalchemy==3.0.5
